\name{coreTDT.results.2.df}
\alias{coreTDT.results.2.df}
\title{
convert coreTDT class to dataframe
}
\description{
convert coreTDT class, i.e., the result list generated inside coreTDT_geneset to dataframe
}
\usage{
coreTDT.results.2.df(coreTDTresults)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{coreTDTresults}{result list generated inside coreTDT_geneset }
}

\value{
dataframe summarized the coreTDT result}
\keyword{ coreTDT }
